Move conditional content into child route, force focus
If a zip code belongs to multiple districts, we display
a simple list of districts to choose from and some instructions.
Previously this was implemented by using if helpers in handlebars
to show the new state after making a network request from a
component.

This isn't a very good API in the first place - it's better
to keep all our network requests in routes and let components
handle UI concerns. However, this also meant that even with
the `{{focused-outlet}}` helper from ember-a11y, screenreaders
had no idea when new content was being rendered - either the
loading spinner or the new information about districts.

The lookup content has been moved to a separate search route
instead of the index route (but it behaves the same since the
search route is mapped to the `/` path). That template now
provides a `{{focusing-outlet}}` where child routes are displayed.
There is a search loading substate, and a dedicated `pick-districts`
child route which shows the information about which districts are
available.

In order to show the loading substate always shows up when a
network request is happening, there is now a `lookup` route which
always redirects — it has no UI of its own. The route picks
the appropriate route to display based on the response (either
display a district, have the user pick a district, or go back to
the main search page if there was an error).

Unfortunately there is an error in the ember-a11y addon which
means that loading substates get correct focus but the finally
rendered route does not. This means the `pick-district` content
is not announced or focused for screenreaders. To force this
content to show up until the issue is resolved, I created a simple
`{{focus-on-render}}` component which grabs focus whenever its first
rendered in the DOM. I wrapped all the `pick-district` content in
this component. I also used the component to ensure messages grab
focus whenever they are added to the DOM.

For more on the ember-a11y issue, see here: https://github.com/ember-a11y/ember-a11y/issues/61

The final result of these changes is this flow:
- go to search page
- submit form without content
- screenreader announces "provide zip" error message
- tab to zip input and add valid zip
- submit form
- screenreader announces "loading"
- screenreader announces pick district information
- tab to district and hit enter
- screenreader announces district information