Make header link use semantic HTML
Instead of leaving the header a regular div with some special
click handling, use a proper `<a>` tag so that the element
gets correct focus and treatment from assitive tech devices.

Instead of doing a route transition, we can do a simple href
to the index URL. Need to also introduce styling to make the
header still maintain the styled look of a title (not a link)
although I did add a hover shadow.